bugfix/mixed model name and path
 The database path was previously being retrieved through the Analysis Services process, specifically from msmdsrv.exe. This approach led to potential confusion when multiple files were being used, as the merging of the file name and the database path could get mixed up.

This commit removes the old implementation of fetching the database path via msmdsrv.exe and adopts the new implementation by @FelixGomes, which leverages the connections present in the Power BI Desktop process to retrieve the database path.

Additionally, the model name and the date it was opened have been added as fields in the dataModel class.